Reworked touch view management.
Previously we relied on getting a touch in the end or cancel phase to
remove the fingertip views. This doesn't always happen, particularly for
swipe to delete on a table view cell, or tap to cancel a swipe to
delete.

Removing all inactive touches after some expiry time seems promising,
but that doesn't work reliably either, because you'll end up removing a
touch for a touch down, wait, then move (because we don't see the
intermediate stationery touches in -sendEvent:)

The hybrid (ugly) solution that we ended up with was removing inactive
touches after expiry only in certain situations, and rely on getting
ended/cancelled touches to remove all others.

This seems to be working well for me in the two applications I tried it
in; we can tune this or rework our approach as necessary.