Alternate fix for overlay window in Fingertips.
If UIKit asks the overlay window for the root view controller, look
through the application window list for a better candidate for that
controller, and return that instead.

From tracing method calls, UIKit returns the top window that 
"_canAffectStatusBarAppearance", apparently checking the bundle of
the class to see if it's an internal window, checking the screen
of the window, and checking the window level.

Attempts to set the window level to all kinds of numbers had failed.

The previous fix, which overrode the setRootViewController method of the
main window, caused problems with touch events not being delivered to
some apps.